fix: move arena functions
